Add files via upload
I updated Person to work with CopyCat, but it can totally be modified further if needed. CopyCat accepts an additional optional parameter and overrides both giveName and whatIDo to give the response of their target instead when possible. Eddie overrides and adds on to askName to add that their real name is technically Eddie <number>.

Descriptions of the classes:
CopyCat: 
Picks a target and then imitates them when possible. Defaults to the person created before them, but accepts an optional parameter to define a different target as well.

Eddie: (named for a friend of mine, because I was stuck for ideas)
Eddie clones. They are all named sequentially as Eddie <number>, but don't let that fool you - as they will make sure to point out, clones are individuals too!